+++ lib/server_connection.ml
@@ -66,6 +66,9 @@ type t =
   ; mutable max_client_stream_id : Stream_identifier.t
   ; mutable max_pushed_stream_id : Stream_identifier.t
   ; mutable receiving_headers_for_stream : Stream_identifier.t option
+        (* Keep track of number of SETTINGS frames that we sent and for which
+         * we haven't eceived an acknowledgment from the client. *)
+  ; mutable unacked_settings : int
   ; mutable did_send_go_away : bool
   ; wakeup_writer : (unit -> unit) ref
         (* From RFC7540§4.3:
@@ -249,12 +252,18 @@ let handle_headers t ~end_stream reqd headers =
    *   limit to be exceeded MUST treat this as a stream error (Section 5.4.2)
    *   of type PROTOCOL_ERROR or REFUSED_STREAM. *)
   if t.current_client_streams + 1 > t.settings.max_concurrent_streams then
-    (* From RFC7540§8.1.4:
-     *   The REFUSED_STREAM error code can be included in a RST_STREAM frame to
-     *   indicate that the stream is being closed prior to any processing
-     *   having occurred. Any request that was sent on the reset stream can be
-     *   safely retried. *)
-    report_stream_error t reqd.Stream.id Error.RefusedStream
+    if t.unacked_settings > 0 then
+      (* From RFC7540§8.1.4:
+       *   The REFUSED_STREAM error code can be included in a RST_STREAM frame to
+       *   indicate that the stream is being closed prior to any processing
+       *   having occurred. Any request that was sent on the reset stream can be
+       *   safely retried.
+       *
+       * Note: if there are pending SETTINGS to acknowledge, assume there was a
+       * race condition and let the client retry. *)
+      report_stream_error t reqd.Stream.id Error.RefusedStream
+    else
+      report_stream_error t reqd.Stream.id Error.ProtocolError
   else
     (* From RFC7540§5.1.2:
      *   Scheduler that are in the "open" state or in either of the "half-closed"
@@ -785,12 +794,18 @@ let process_rst_stream_frame t { Frame.frame_header; _ } error_code =
 
 let process_settings_frame t { Frame.frame_header; _ } settings =
   let open Scheduler in
-  (* TODO: what to do in the case of receiving an acked SETTINGS frame for
-   * settings we didn't send? *)
   let { Frame.flags; _ } = frame_header in
   (* We already checked that an acked SETTINGS is empty. Don't need to do
    * anything else in that case *)
-  if not Flags.(test_ack flags) then
+  if Flags.(test_ack flags) then (
+    t.unacked_settings <- t.unacked_settings - 1;
+    if t.unacked_settings < 0 then
+      (* The server is ACKing a SETTINGS frame that we didn't send *)
+      let additional_debug_data =
+        "Received SETTINGS with ACK but no ACK was pending"
+      in
+      report_connection_error t ~additional_debug_data Error.ProtocolError)
+  else
     match Settings.check_settings_list settings with
     | None ->
       (* From RFC7540§6.5:
@@ -873,6 +888,7 @@ let process_settings_frame t { Frame.frame_header; _ } settings =
        *   ACK (0x1): [...] When this bit is set, the payload of the SETTINGS
        *   frame MUST be empty. *)
       Writer.write_settings t.writer frame_info [];
+      t.unacked_settings <- t.unacked_settings + 1;
       wakeup_writer t
     | Some error ->
       handle_error t error
@@ -1153,6 +1169,7 @@ let create
       ; max_pushed_stream_id = 0l
       ; receiving_headers_for_stream = None
       ; did_send_go_away = false
+      ; unacked_settings = 0
       ; wakeup_writer = ref default_wakeup_writer
       ; hpack_encoder = Hpack.Encoder.(create settings.header_table_size)
       ; hpack_decoder = Hpack.Decoder.(create settings.header_table_size)