+++ CHANGES.md
@@ -13,6 +13,8 @@ Unreleased
   ([#16](https://github.com/anmonteiro/ocaml-h2/pull/16))
 - h2: Remove dependency on the `result` package
   ([#18](https://github.com/anmonteiro/ocaml-h2/pull/18))
+- h2: Track SETTINGS frames that haven't been acknowledged by the peer
+  ([#22](https://github.com/anmonteiro/ocaml-h2/pull/22))
 
 0.1.0 2019-03-27
 --------------